@@ -6,7 +6,7 @@ plugins {
 }
 
 android {
-    compileSdk 33
+    compileSdk 34
 
     Properties properties = new Properties()
     properties.load(project.rootProject.file('local.properties').newDataInputStream())
@@ -14,11 +14,10 @@ android {
     defaultConfig {
         applicationId "com.dongyang.android.youdongknowme"
         minSdk 24
-        targetSdk 33
-        versionCode 10
-        versionName "2.0.0"
+        targetSdk 34
+        versionCode 13
+        versionName "2.1.0"
 
-        buildConfigField "String", "API_KEY", properties['API_KEY']
         buildConfigField "String", "BASE_URL", properties['BASE_URL']
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
